backend/requirements.txt
fastapi
uvicorn
pandas
requests
python-dotenv
entsoe-py
geopy
